SecProj/S3_LogAnalyzer/requirements.txt
bcrypt==3.1.7
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.3.2
distlib==0.3.9
docopt==0.6.2
enum34==1.1.10
filelock==3.2.1
idna==2.10
importlib-metadata==2.1.3
importlib-resources==3.3.1
ipaddress==1.0.23
numpy==1.16.6
pandas==0.24.2
paramiko==2.12.0
pathlib2==2.3.7.post1
pipreqs==0.4.11
platformdirs==2.0.2
pycparser==2.21
PyNaCl==1.4.0
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2024.2
requests==2.27.1
scandir==1.10.0
singledispatch==3.7.0
six==1.17.0
typing==3.10.0.0
urllib3==1.26.20
virtualenv==20.15.1
yarg==0.1.10
zipp==1.2.0
